% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choc.R
\name{choc}
\alias{choc}
\title{carry out a choc analysis on a multivariate time series}
\usage{
choc(mydata, H, timevar, resolution = 100)
}
\arguments{
\item{mydata}{a data frame or matrix with one column per time series and one row per observation (event). Marginal distributions are assumed to follow a normal distribution}

\item{H}{either a function from library ks to estimate a bandwith matrix, or directly a bandwith matrix. Exemples of function \code{\link[ks]{Hpi}} or \code{\link[ks]{Hscv}}...}

\item{timevar}{a vector specifying the time step for each observation. Several observations per time step are required. Observations of a same time step are assumed to be replicates.}

\item{resolution}{grid resolution on which the densities of probability will be computed. The resolution does not affect the result: high resolution increases the resolution of final diagrams but increases computation time and memory usage. The number of observations per time step should be roughly similar}
}
\value{
a list with two elements: \enumerate{
\item the list of fitted kernels per time step
\item a dafaframe with one row per time series and a column tau that gives the tau of Kendall trend test
}
}
\description{
carry out a choc analysis on a multivariate time series
}
\examples{
#generate artificial data set
#two time series measured on 40 time steps with 365 observations per time step.
#the two series follow a multinormal time series with a tend on means and a constant
#covariance matrix
library(MASS)
library(ks)
tvar <- rep(1:40,times=100) #times steps
meansX <-tvar/40 #trend on 1st variable
meansY <- -0.5*tvar/40 #trend on 2nd variable
sigma <- matrix(c(1,.1,.1,1),2,2) #covariance matrix
values <- t(apply(cbind(meansX,meansY),1,function(mu) mvrnorm(1,mu,sigma))) #generate the values
H <- Hpi #choose the default bandwith
res <- choc(values,H,tvar)

}
