% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chocIvlev.R
\name{chocIvlev}
\alias{chocIvlev}
\title{chocIvlev
carry out a choc analysis by comparing a chocR of realised niche and a chocR
of available niche}
\usage{
chocIvlev(chocRealised, chocAvailable)
}
\arguments{
\item{chocRealised}{a chocR object of realised niche \link{choc}}

\item{chocAvailable}{a chocR object of available niche consistent \link{choc}
with chocRealised}
}
\value{
a list with 3 elements: \enumerate{
\item chocRealised
\item chocAvailable
\item list_ivlev the list of ivelev per time step computed on grid
\item grid the Kendall tau on the grid
}
}
\description{
chocIvlev
carry out a choc analysis by comparing a chocR of realised niche and a chocR
of available niche
}
