% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_confidenceIvlev.R
\name{estimate_confidenceIvlev}
\alias{estimate_confidenceIvlev}
\title{estimate_confidenceIvlev
estimate confidence intervals for chocIvlev analysis}
\usage{
estimate_confidenceIvlev(
  mychocIvlev,
  conf = 0.95,
  nb_replicates = 500,
  ncores = 1,
  progressbar = TRUE
)
}
\arguments{
\item{mychocIvlev}{an object returned by \link{chocIvlev}}

\item{conf}{size of the confidence interval}

\item{nb_replicates}{number of replicates used to assess confidence intervals}

\item{ncores}{Number of cores used. The parallelization will take place only if OpenMP is supported (default 1)}

\item{progressbar}{(default TRUE) show progressbar (might be a bit slower)}
}
\value{
an updated version of mychoc with two columns added to mychoc$grid which corresponds to the bounds of the confidence interval
}
\description{
estimate_confidenceIvlev
estimate confidence intervals for chocIvlev analysis
}
