% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallIvlev.R
\name{overallIvlev}
\alias{overallIvlev}
\title{overallIvlev
computes the average preferenum over the whole study period}
\usage{
overallIvlev(chocRealised, chocAvailable)
}
\arguments{
\item{chocRealised}{a chocR object of realised niche \link{choc}}

\item{chocAvailable}{a chocR object of available niche consistent \link{choc}
with chocRealised}
}
\value{
a data.frame with the grid and an ivlev index
}
\description{
overallIvlev
computes the average preferenum over the whole study period
}
