% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallNiche.R
\name{overallNiche}
\alias{overallNiche}
\title{overallNiche
computes the average niche over the whole study period}
\usage{
overallNiche(mychoc)
}
\arguments{
\item{mychoc}{a chocR object of realised niche \link{choc}
with chocRealised}
}
\value{
a data.frame with the grid and and the corresponding density
}
\description{
overallNiche
computes the average niche over the whole study period
}
