% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choc.R
\name{plot_choc}
\alias{plot_choc}
\title{Plot the result of choc analysis
red means that there is a significant decrease of occurence of the association, pale red stands for non significant trend
pale green for non significant positive trend and green for significant positive trend}
\usage{
plot_choc(mychoc, ivar = c(1, 2))
}
\arguments{
\item{mychoc}{a list returned by \code{\link{estimate_confidence}}}

\item{ivar}{the two variables that should be plotted}
}
\value{
a ggplot object
}
\description{
Plot the result of choc analysis
red means that there is a significant decrease of occurence of the association, pale red stands for non significant trend
pale green for non significant positive trend and green for significant positive trend
}
\examples{
#generate artificial data set
#two time series measured on 40 time steps with 365 observations per time step.
#the two series follow a multinormal time series with a tend on means and a constant
#covariance matrix
library(MASS)
library(ks)
tvar <- rep(1:40,times=100) #times steps
meansX <- tvar/40 #trend on 1st variable
meansY <- -0.5*tvar/40 #trend on 2nd variable
sigma <- matrix(c(1,.1,.1,1),2,2) #covariance matrix
values <- t(apply(cbind(meansX,meansY),1,function(mu) mvrnorm(1,mu,sigma))) #generate the values
H<-Hpi #choose the default bandwith
res <- choc(values,H,tvar)
res_confid <- estimate_confidence(res,"perm",0.95,50) #here we put a low number of replicates to limit computation time
g<-plot_choc(res_condit)
print(g)

}
