% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_choc}
\alias{res_choc}
\title{sample results of a \code{\link{choc}} function}
\format{
see \code{\link{choc}}
}
\usage{
res_choc
}
\description{
sample results of a \code{\link{choc}} function
}
\keyword{datasets}
