% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_confid}
\alias{res_confid}
\title{sample results of a \code{\link{estimate_confidence}} function}
\format{see \code{\link{estimate_confidence}}}
\usage{
res_confid
}
\description{
sample results of a \code{\link{estimate_confidence}} function
}
\keyword{datasets}
