% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choc.R
\name{plot_choc}
\alias{plot_choc}
\title{plot_choc
Plot the result of choc analysis}
\usage{
plot_choc(mychoc, ivar = c(1, 2))
}
\arguments{
\item{mychoc}{a list returned by \code{\link{estimate_confidence}}}

\item{ivar}{the two variables that should be plotted}
}
\value{
a ggplot object
}
\description{
plot_choc
Plot the result of choc analysis
}
\section{Details}{

red means that there is a significant decrease of occurence of the association, pale red stands for non significant trend
pale green for non significant positive trend and green for significant positive trend
}

\examples{
#example of results provided by estimate_confidence
data(res_confid)
g<-plot_choc(res_confid)
print(g)

}
